Move path from mathswe-client/origin to mathswe-client/req/origin
